update: 添加问题“3195.包含所有1的最小矩形面积I”的代码和题解 (#1093)
* archive: many days(words+837题解小改)


* 3195: WA.cpp (#1092) + docs(jp) + init

* 3195: AC.cpp+WA.py (#1092) - 和CPP的WA原因一模一样

cpp - AC,64.89%,94.68%

* 3195: AC.py+rs + LEETCODE - bug(#1092) - go写好快

py - AC,25.00%,81.25%
rust - AC,40.00%,40.00%
Compile timed out. Please try again. If the issue persists, please report it to us.

* 3195: AC.go+CE.java (#1092)

go - AC,68.42%,42.10%

* 3195: AC.java (#1092) - java也写好快

java - AC,98.08%,84.62%

* update: 添加问题“3195.包含所有1的最小矩形面积I”的代码和题解 (#1093)


* fix(words): (9个月前)拼写错误的日语单词

Update Solutions/Other-Japanese-LearningNotes.md


* typo: grid

Update Solutions/LeetCode 3195.包含所有1的最小矩形面积I.md


* typo: 即为

Update Solutions/LeetCode 3195.包含所有1的最小矩形面积I.md


---------

+++ Codes/0837-new-21-game.rs
@@ -2,7 +2,7 @@
  * @Author: LetMeFly
  * @Date: 2025-08-17 19:33:11
  * @LastEditors: LetMeFly.xyz
- * @LastEditTime: 2025-08-17 22:31:00
+ * @LastEditTime: 2025-08-17 22:55:04
  */
 impl Solution {
     pub fn new21_game(n: i32, k: i32, max_pts: i32) -> f64 {
@@ -10,7 +10,7 @@ impl Solution {
         let max_pts: usize = max_pts as usize;
         let n: usize = n as usize;
 
-        let mut dp: Vec<f64> = vec![0 as f64; k + max_pts];
+        let mut dp: Vec<f64> = vec![0.; k + max_pts];
         let mut s: f64 = 0.;
         for i in k..(k+max_pts) {
             if i <= n {
